+++ Assets/Scripts/NPC/FinalBoss.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5f1aea43799a7064f8b339ceed8d0ca7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
