+++ Assets/Scripts/NPC/FinalBoss/BossWalk.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4ee0724a9076dfd4d84c9cbddc65aaf5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
