+++ Assets/Scripts/NPC/FinalBoss/JumpAttack.cs
@@ -1,0 +1,40 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class JumpAttack : StateMachineBehaviour
+{
+    private Rigidbody2D rb;
+    private FinalBoss FBoss;
+    // OnStateEnter is called when a transition starts and the state machine starts to evaluate this state
+    override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+    {
+       FBoss=animator.GetComponent<FinalBoss>();
+       rb=FBoss.rb;
+       FBoss.LockPlayer();
+    }
+
+    // OnStateUpdate is called on each Update frame between OnStateEnter and OnStateExit callbacks
+    override public void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+    {
+       rb.velocity=new Vector2(rb.velocityX,5);
+    }
+
+    // OnStateExit is called when a transition ends and the state machine finishes evaluating this state
+    /*override public void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+    {
+       animator.SetBool("Fall",true);
+    }*/
+
+    // OnStateMove is called right after Animator.OnAnimatorMove()
+    //override public void OnStateMove(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+    //{
+    //    // Implement code that processes and affects root motion
+    //}
+
+    // OnStateIK is called right after Animator.OnAnimatorIK()
+    //override public void OnStateIK(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+    //{
+    //    // Implement code that sets up animation IK (inverse kinematics)
+    //}
+}
