+++ Assets/Scripts/NPC/FinalBoss/JumpAttack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b57b493c82493514eb3e07dab214ae0c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
