+++ ProjectSettings/TagManager.asset
@@ -22,7 +22,7 @@ TagManager:
   - Rats
   - Jaguar
   - Owl
-  - 
+  - Boss
   - 
   - 
   -
